+++ packages/lumx-react/src/components/badge/Badge.stories.tsx
@@ -1,23 +1,23 @@
 import { mdiHeart } from '@lumx/icons';
 import { AspectRatio, Badge, ColorPalette, Icon, Size, Thumbnail, ThumbnailVariant } from '@lumx/react';
 import { select, text } from '@storybook/addon-knobs';
-import React from 'react';
+import React, { Fragment } from 'react';
 
 export default { title: 'LumX components/badge/Badge' };
 
-export const simpleBadgeWithValue = () => (
+export const WithText = () => (
     <Badge color={select('Colors', ColorPalette, ColorPalette.blue)}>
         <span>{text('Value', '30')}</span>
     </Badge>
 );
 
-export const simpleBadgeWithIcon = () => (
+export const WithIcon = () => (
     <Badge color={select('Colors', ColorPalette, ColorPalette.red)}>
         <Icon icon={mdiHeart} />
     </Badge>
 );
 
-export const simpleBadgeWithThumbnail = () => (
+export const WithThumbnail = () => (
     <Badge color={select('Colors', ColorPalette, ColorPalette.light)}>
         <Thumbnail
             alt="Logo"
@@ -27,4 +27,19 @@ export const simpleBadgeWithThumbnail = () => (
             variant={ThumbnailVariant.rounded}
         />
     </Badge>
+);
+
+export const AllColors = () => (
+    <dl>
+        {Object.values(ColorPalette).map((color) => (
+            <Fragment key={color}>
+                <dt>{color}</dt>
+                <dd>
+                    <Badge color={color}>
+                        <Icon icon={mdiHeart} />
+                    </Badge>
+                </dd>
+            </Fragment>
+        ))}
+    </dl>
 );