+++ packages/lumx-react/src/stories/knobs/sizeKnob.ts
@@ -1,5 +1,5 @@
 import { Size } from '@lumx/react';
 import { enumKnob } from '@lumx/react/stories/knobs/enumKnob';
 
-export const sizeKnob = (size?: Size) =>
-    enumKnob('Size', [undefined, Size.xxs, Size.xs, Size.s, Size.m, Size.l, Size.xl, Size.xxl] as const, size as any);
+export const sizeKnob = (name = 'Size', size?: Size) =>
+    enumKnob(name, [undefined, Size.xxs, Size.xs, Size.s, Size.m, Size.l, Size.xl, Size.xxl] as const, size as any);